requirements.txt
# GitHub API
PyGithub==1.55

# Web scraping
requests==2.26.0
beautifulsoup4==4.10.0

# PDF processing
PyPDF2==1.26.0

# Natural Language Processing
nltk==3.6.5

# Vector Embedding
openai==1.3.0

# Vector Database
pinecone-client==2.2.4

# Other utilities
numpy==1.26.4

# Environment variables
python-dotenv==0.19.1

# Local vector store
faiss-cpu==1.7.4